+++ src/osu_2015/stars.rs
@@ -77,7 +77,7 @@ pub fn stars(map: &Beatmap, mods: u32, passed_objects: Option<usize>) -> OsuDiff
         scaling_factor,
     );
 
-    while h.base.time as f32 > current_section_end {
+    while h.base.time > current_section_end {
         current_section_end += section_len;
     }
 
@@ -95,7 +95,7 @@ pub fn stars(map: &Beatmap, mods: u32, passed_objects: Option<usize>) -> OsuDiff
             scaling_factor,
         );
 
-        while h.base.time as f32 > current_section_end {
+        while h.base.time > current_section_end {
             aim.save_current_peak();
             aim.start_new_section_from(current_section_end);
             speed.save_current_peak();
